Move `Agent` and `message::Outbound` to `stateful` :tada:
