+++ packages/engine/lib/stateful/Cargo.toml
@@ -14,3 +14,4 @@ serde_json = "1.0.59"
 serde-aux = "0.6.1"
 thiserror = "1.0.21"
 tracing = "0.1.29"
+uuid = { version = "0.8.1", features = ["v4", "serde"] }
