+++ packages/engine/lib/stateful/src/error.rs
@@ -20,6 +20,12 @@ pub enum Error {
 
     #[error("Couldn't acquire exclusive lock on object")]
     ProxyExclusiveLock,
+
+    #[error("Serde Error: {0}")]
+    Serde(#[from] serde_json::Error),
+
+    #[error("Could not parse outbound message: {0}")]
+    OutboundMessageParse(#[from] crate::message::OutboundError),
 }
 
 impl From<&str> for Error {
