+++ packages/engine/src/datastore/test_utils.rs
@@ -3,7 +3,7 @@ use std::sync::Arc;
 use rand::{prelude::StdRng, Rng, SeedableRng};
 use serde::{Deserialize, Serialize};
 use stateful::{
-    agent::{AgentSchema, AgentStateField},
+    agent::{Agent, AgentSchema, AgentStateField},
     field::{
         FieldScope, FieldSpec, FieldSpecMap, FieldType, FieldTypeVariant, RootFieldSpec,
         RootFieldSpecCreator,
@@ -20,7 +20,6 @@ use crate::{
         error::Error,
         schema::{last_state_index_key, EngineComponent},
     },
-    hash_types::state::Agent,
     proto::{ExperimentRunBase, InitialState, InitialStateName, ProjectBase},
     simulation::package::creator::{get_base_agent_fields, PackageCreators},
 };
