+++ packages/engine/src/simulation/package/init/mod.rs
@@ -6,12 +6,14 @@ use std::sync::Arc;
 
 use async_trait::async_trait;
 pub use packages::{InitTask, InitTaskMessage, Name, PACKAGE_CREATORS};
-use stateful::field::{FieldSpecMapAccessor, RootFieldSpec, RootFieldSpecCreator};
+use stateful::{
+    agent::Agent,
+    field::{FieldSpecMapAccessor, RootFieldSpec, RootFieldSpecCreator},
+};
 
 use crate::{
     config::{ExperimentConfig, Globals, SimRunConfig},
     datastore::schema::EngineComponent,
-    hash_types::Agent,
     simulation::{
         comms::package::PackageComms,
         package::{
