+++ packages/engine/src/simulation/package/init/packages/js_py/mod.rs
@@ -6,6 +6,7 @@ use std::{
 use async_trait::async_trait;
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
+use stateful::agent::Agent;
 
 use crate::{
     config::ExperimentConfig,
@@ -15,7 +16,7 @@ use crate::{
         enum_dispatch::{enum_dispatch, RegisterWithoutTrait, TaskSharedStore},
         package::init::{
             packages::js_py::{js::JsInitTask, py::PyInitTask},
-            Agent, Arc, FieldSpecMapAccessor, GetWorkerExpStartMsg, GetWorkerSimStartMsg, InitTask,
+            Arc, FieldSpecMapAccessor, GetWorkerExpStartMsg, GetWorkerSimStartMsg, InitTask,
             InitTaskMessage, MaybeCpuBound, Package as InitPackage, PackageComms, PackageCreator,
             SimRunConfig,
         },
