+++ packages/engine/src/simulation/package/init/packages/json/mod.rs
@@ -1,12 +1,13 @@
 use async_trait::async_trait;
 use serde_json::Value;
+use stateful::agent::Agent;
 
 use crate::{
     datastore::schema::EngineComponent,
     proto::{ExperimentRunTrait, InitialStateName},
     simulation::{
         package::init::{
-            Agent, Arc, ExperimentConfig, FieldSpecMapAccessor, GetWorkerExpStartMsg,
+            Arc, ExperimentConfig, FieldSpecMapAccessor, GetWorkerExpStartMsg,
             GetWorkerSimStartMsg, MaybeCpuBound, Package as InitPackage, PackageComms,
             PackageCreator, SimRunConfig,
         },
